WIP: tests for centoid, model, orietation
