+++ tests/position/test_dlc_model.py
@@ -1,3 +1,6 @@
+import pytest
+
+
 def test_model_params_default(sgp):
     assert sgp.v1.DLCModelParams.get_default() == {
         "dlc_model_params_name": "default",
@@ -8,3 +11,8 @@ def test_model_params_default(sgp):
             "model_prefix": "",
         },
     }
+
+
+def test_model_input_assert(sgp):
+    with pytest.raises(AssertionError):
+        sgp.v1.DLCModelInput().insert1({"config_path": "/fake/path/"})
