+++ tests/position/test_dlc_pos.py
@@ -1,4 +1,3 @@
-import pandas as pd
 import pytest
 from numpy import isclose as np_isclose
